Set up VendingMachine with first passing tests: select_product
It allows to initialise a VendingMachine with a list of Product objects 
and call #select_product. The product is selected by inserting the name 
of the desired product.

Tests include checking the instance where a product does not exist and 
not enough money is inserted.

The Product object is not yet implemented; the tests are run by stubs of 
a Product object.