Initialise CashHolder object with optional input_value
The CashHolder will contain and handle any logic to do with money; this 
includes the  storing of cash but eventually it will also do the  
calculations of what kind of change to return if any.

If a value is given on initialisation, that's the starting amount of 
money a vending machine has. If no value is given, it starts with 0 
money in the machine.