CashHolder accepts both the item_price and given_funds, and returns the  change between the two.
Instead of the machine taking in the full amount of money given, we now 
return change if more money is given than the price of the product. This 
does not yet take into consideration the different coins.

The responsibility whether enough money is given has now also moved from 
the VendingMachine to the CashHolder. This makes sense, as the 
VendingMachine itself shouldn't care about any money handling so it's 
more in line with the single responsibility principle.