Change return of #store_funds to not return change as a value but an array of coins
The instructions state we need to return change `in denominations of:
1p, 2p, 5p, 10p, 20p, 50p, £1, £2`. The private #return_change method
loops through each coin, highest to lowest, and sorts out what coins can
be returned. We always attempt to return as big as possible coins.

The function can likely be refactored, which may be addressed in a
future commit.

The test has been updated to check for the correct array to be returned