CashHolder to handle input and storing of coin instances instead of total amounts
Instead of a user passing in a total amount value into the vending 
machine, we now handle separate coin instances they pass. This allows to 
also return certain coins as change if they are available. 

If they are not available, we _should_ return the nearerst possible 
amount (rounded down), this needs to be confirmed in a separate unit 
test.