Add in forgotten #get_total_value method for CashHolder
When moving away from the cash holder thinking in #current_value and 
moving to coins, we no longer have the option to see what the total 
amount of money is inside the CashHolder. 

This method brings that option back in for the vending machine holder to 
see, as it's one of the user stories requirements.